requirements.txt
fpdf==1.7.2
geojson==2.5.0
mrcnn==0.2
nose==1.3.7
pyshp==2.3.0
tksheet==5.0.34
traitlets==5.1.0
pdal~=2.4.1
numpy==1.22.4
Pillow==9.1.1
pyproj~=3.3.1
rasterio~=1.2.10
SQLAlchemy~=1.4.37
vispy~=0.10.0
geopandas~=0.10.2
zlib~=1.2.12
scikit-learn~=1.1.1
opencv-python==4.3.0.36